refactor(home view): section specs (#6133)
* test: ActiveBids

* test: AuctionLotsForYou

* test: Auctions

* test: CuratorsPicksEmerging

* test: DiscoverSomethingNew (stub)

* fix: error message when not displayable

* test: ExploreByCategory

* test: FeaturedFairs

* test: HeroUnits

* test: LatestActivity

* test: LatestArticles

* test: LatestAuctionResults

* test: MarketingCollections

* test: News

* test: NewWorksForYou

* test: NewWorksFromGalleriesYouFollow

* test: RecentlyViewedArtworks

* test: RecommendedArtists

* test: RecommendedArtworks

* test: SimilarToRecentlyViewedArtworks

* test: Tasks

* test: GalleriesNearYou

This should have been deleted in #6116, since the
test was copied to the new location already

* test: TrendingArtists

* test: ViewingRooms

* test: ShowsForYou

* docs: update section docs with testing conventions
