+++ src/schema/v2/homeView/sections/__tests__/DiscoverSomethingNew.test.ts
@@ -1,0 +1,3 @@
+describe("DiscoverSomethingNew", () => {
+  it.skip("needs specs", () => {})
+})
